+++ src/main/java/org/accounting/system/endpoints/ProviderEndpoint.java
@@ -7,6 +7,7 @@ import jakarta.validation.Valid;
 import jakarta.validation.constraints.Max;
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotNull;
+import jakarta.ws.rs.BadRequestException;
 import jakarta.ws.rs.Consumes;
 import jakarta.ws.rs.DELETE;
 import jakarta.ws.rs.DefaultValue;
@@ -27,6 +28,7 @@ import org.accounting.system.dtos.pagination.PageResource;
 import org.accounting.system.dtos.provider.ProviderRequestDto;
 import org.accounting.system.dtos.provider.ProviderResponseDto;
 import org.accounting.system.dtos.provider.UpdateProviderRequestDto;
+import org.accounting.system.entities.HierarchicalRelation;
 import org.accounting.system.entities.projections.ProviderProjectionWithProjectInfo;
 import org.accounting.system.enums.Collection;
 import org.accounting.system.enums.Operation;
@@ -186,6 +188,11 @@ public class ProviderEndpoint {
     public Response save(@Valid @NotNull(message = "The request body is empty.") ProviderRequestDto providerRequestDto, @Context UriInfo uriInfo) {
 
         var serverInfo = new AccountingUriInfo(serverUrl.concat(basePath).concat(uriInfo.getPath()));
+
+        if(providerRequestDto.id.contains(HierarchicalRelation.PATH_SEPARATOR)){
+
+            throw new BadRequestException("Provider ID should not contain a dot character.");
+        }
 
         providerService.existById(providerRequestDto.id);
         providerService.existByName(providerRequestDto.name);